requirement.txt
Flask==2.3.4
python-dotenv==1.0.1
requests==2.31.0
pyzbar==0.1.9
Pillow==10.1.0
